Fixed build issue, moved types outside pages
